Исправляет ошибку повторной отрисовки
